pyproject.toml
[project]
name = "bcms-ocr"
version = "0.1.0"
description = "Intel Arc friendly OCR pipeline with business card field extraction."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers==4.46.3",
    "Pillow>=10.0.0",
    "torchvision>=0.16.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "addict>=2.4.0",
    "requests>=2.31.0",
    "easyocr>=1.7.1",
]

[project.optional-dependencies]
intel = ["intel-extension-for-pytorch>=2.1.0"]

[project.scripts]
ocr-local = "ocr_local:main"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
include = ["ocr_local.py"]
